Dissolve handshake module into host/guest modules
The most natural seam for modules isn't CLI -> Handshake -> Zellij
layers anymore, so I've broken it up. ALPN seems like a natural part of
protocol.rs, so it goes there. init_endpoint() moves up to main where
it's actually used. Whole-binary tests go to tests.rs, which is just a
child of main.rs.
